Resticting Post

+++ app/Livewire/Candidate/CandidateDelete.php
@@ -3,6 +3,7 @@
 namespace App\Livewire\Candidate;
 
 use App\Models\Candidate;
+use App\Models\Votes;
 use App\Traits\EncryptDecrypt;
 use Illuminate\View\View;
 use Livewire\Attributes\Locked;
@@ -33,7 +34,17 @@ class CandidateDelete extends Component
 
     public function deleteCandidate(): void
     {
-        $delete = Candidate::where('id',$this->decryptId($this->id))->delete();
+        //dd($this->decryptId($this->id));
+        $candidate_check = Votes::where('candidate_id', $this->decryptId($this->id))->count();
+        if ($candidate_check > 0) {
+            $this->dispatch(
+                'notify',
+                title: 'info',
+                timer: 9000,
+                message: 'Sorry, '. $this->name. ' is already running for election');
+
+        }else{
+            $delete = Candidate::where('id',$this->decryptId($this->id))->delete();
 
         ($delete)
             ? $this->dispatch('notify', title: 'success', message:  ' '.$this->name. ' Deleted successfully')
@@ -41,7 +52,9 @@ class CandidateDelete extends Component
 
         $this->dispatch('dispatch-candidate-deleted')->to(CandidateTable::class);
 
+        }
         $this->DeleteCandidateModal  = false;
+
 
     }
