+++ app/Livewire/Posts/PostDelete.php
@@ -3,6 +3,7 @@
 namespace App\Livewire\Posts;
 
 use App\Models\Post;
+use App\Models\Votes;
 use App\Traits\EncryptDecrypt;
 use Illuminate\View\View;
 use Livewire\Attributes\Locked;
@@ -33,13 +34,23 @@ class PostDelete extends Component
 
      public function deletePost(): void
     {
-        $delete = Post::destroy($this->decryptId($this->id));
+        $candidate_check = Votes::where('post_id', $this->decryptId($this->id))->count();
+        if ($candidate_check > 0) {
+            $this->dispatch(
+                'notify',
+                title: 'info',
+                timer: 9000,
+                message: 'Sorry, '. $this->name. ' is active for this current election campaign');
+
+        }else{
+            $delete = Post::destroy($this->decryptId($this->id));
 
         ($delete)
             ? $this->dispatch('notify', title: 'success', message:  ' '.$this->name. ' Deleted successfully')
             : $this->dispatch('notify', title: 'fail', message: 'Ops!! Something went wrong');
 
         $this->dispatch('dispatch-post-deleted')->to(PostTable::class);
+        }
 
         $this->DeletePostModal  = false;
